Refactor reducers’s file names under Users screen
Replace singulars with plurals.
