+++ src/screens/Users/reducers.js
@@ -2,9 +2,9 @@
 import { combineReducers } from 'redux';
 
 // Reducers
-import profileReducer from './screens/Profile/reducer';
-import signinReducer from './screens/SignIn/reducer';
-import signupReducer from './screens/SignUp/reducer';
+import profileReducer from './screens/Profile/reducers';
+import signinReducer from './screens/SignIn/reducers';
+import signupReducer from './screens/SignUp/reducers';
 
 // Combine reducers
 export default combineReducers({
