+++ src/screens/reducers.js
@@ -3,7 +3,7 @@ import { combineReducers } from 'redux';
 
 // Reducers
 import surveysReducer from './Surveys/reducer';
-import usersReducer from './Users/reducer';
+import usersReducer from './Users/reducers';
 
 // Combine reducers
 export default combineReducers({
